Rename the button.js file to index.js to comply with our naming convention
