+++ src/components/screens/bookmarks/addBookmark/addBookmark.js
@@ -7,7 +7,7 @@ import BoxHeader from '../../../toolbox/box/header';
 import BoxContent from '../../../toolbox/box/content';
 import BoxFooter from '../../../toolbox/box/footer';
 import { Input } from '../../../toolbox/inputs';
-import { PrimaryButton, SecondaryButton } from '../../../toolbox/buttons/button';
+import { PrimaryButton, SecondaryButton } from '../../../toolbox/buttons';
 import styles from './addBookmark.css';
 import { getIndexOfBookmark } from '../../../../utils/bookmarks';
 import { tokenMap } from '../../../../constants/tokens';
