+++ src/components/screens/hwWalletLogin/requestPin.js
@@ -1,7 +1,7 @@
 import { to } from 'await-to-js';
 import React from 'react';
 import { Input } from '../../toolbox/inputs';
-import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons';
 import { getPublicKey, validatePin } from '../../../utils/hwManager';
 import externalLinks from '../../../constants/externalLinks';
 import styles from './requestPin.css';
