+++ src/components/screens/hwWalletLogin/selectAccount/selectAccount.js
@@ -1,7 +1,7 @@
 import { to } from 'await-to-js';
 import React from 'react';
 import { toast } from 'react-toastify';
-import { TertiaryButton } from '../../../toolbox/buttons/button';
+import { TertiaryButton } from '../../../toolbox/buttons';
 import { getAccountsFromDevice } from '../../../../utils/hwManager';
 import AccountCard from './accountCard';
 import LoadingIcon from '../loadingIcon';
