+++ src/components/screens/register/accountCreated.js
@@ -3,7 +3,7 @@ import { withTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import routes from '../../../constants/routes';
-import { PrimaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton } from '../../toolbox/buttons';
 import registerStyles from './register.css';
 import styles from './accountCreated.css';
 import Illustration from '../../toolbox/illustration';
