+++ src/components/screens/register/backupPassphrase.js
@@ -1,7 +1,7 @@
 import { withTranslation } from 'react-i18next';
 import React from 'react';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
-import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons';
 import PassphraseBackup from '../../shared/passphraseBackup';
 import registerStyles from './register.css';
