+++ src/components/screens/register/chooseAvatar.js
@@ -3,7 +3,7 @@ import { withTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import routes from '../../../constants/routes';
-import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons';
 import AccountVisual from '../../toolbox/accountVisual';
 import registerStyles from './register.css';
 import styles from './chooseAvatar.css';
