+++ src/components/shared/analyticsDialog/analyticsDialog.js
@@ -4,7 +4,7 @@ import { toast } from 'react-toastify';
 import PropTypes from 'prop-types';
 import Dialog from '../../toolbox/dialog/dialog';
 import FlashMessageHolder from '../../toolbox/flashMessage/holder';
-import { PrimaryButton, SecondaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton, SecondaryButton } from '../../toolbox/buttons';
 import externalLinks from '../../../constants/externalLinks';
 import styles from './analyticsDialog.css';
