+++ src/components/shared/loadLatestButton/index.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { PrimaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton } from '../../toolbox/buttons';
 import Icon from '../../toolbox/icon';
 import useServiceSocketUpdates from '../../../hooks/useServiceSocketUpdates';
 import styles from './loadLatestButton.css';
