+++ src/components/shared/newReleaseMessage/newReleaseMessage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { withTranslation } from 'react-i18next';
 import PropTypes from 'prop-types';
 import FlashMessage from '../../toolbox/flashMessage/flashMessage';
-import { TertiaryButton } from '../../toolbox/buttons/button';
+import { TertiaryButton } from '../../toolbox/buttons';
 import styles from './newReleaseMessage.css';
 
 const NewReleaseMessage = ({
