+++ src/components/shared/notFound/index.js
@@ -2,7 +2,7 @@ import { Link } from 'react-router-dom';
 import React from 'react';
 import { withTranslation } from 'react-i18next';
 import styles from './notFound.css';
-import { PrimaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton } from '../../toolbox/buttons';
 import Illustration from '../../toolbox/illustration';
 import routes from '../../../constants/routes';
