+++ src/components/shared/transactionResult/index.js
@@ -1,6 +1,6 @@
 /* eslint-disable complexity */
 import React from 'react';
-import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons/button';
+import { PrimaryButton, TertiaryButton } from '../../toolbox/buttons';
 import { getErrorReportMailto } from '../../../utils/helpers';
 import styles from './transactionResult.css';
 import Illustration from '../../toolbox/illustration';
